extracted some methods in FilesHelper.kt object, code cleanup
